\name{CRFCSD-package}
\alias{CRFCSD-package}
\alias{CRFCSD}
\docType{package}
\title{
\packageTitle{CRFCSD}
}
\description{
%\packageDescription{CRFMCCS}
A methodology to estimate the parameters for the cure rate frailty
models with clustered current status data.
}

\author{
\packageAuthor{CRFCSD}

Maintainer:\packageMaintainer{CRFCSD}
}

%\references{
%Wang, T., He, K., Ma, W., Bandyopadhyay, D. and Sinha S. Mixture Cure frailty Generalized Odds Ratio Model for Clustered Current Status Data. Preprint. 
%}
